modif gestion erreur auth

+++ services/users/auth/authController.js
@@ -1,5 +1,5 @@
 import { app } from '../../gateway/server.js';
-import { user_db } from '../usersServer.js';
+import { user_db, httpError } from '../usersServer.js';
 import { authenticator } from 'otplib';
 
 export const authRegister = async function (req, reply) {
@@ -26,8 +26,12 @@ export const authRegister = async function (req, reply) {
 		
 		return (reply.code(201).send("New entry in database"));
 	} catch (err) {
-		const e = new Error(err.message);
-		e.statusCode = 409;
+		console.error(`\nERROR authRegister: ${err.message}\n`);
+		const e = new Error();
+		if (err.code === "SQLITE_CONSTRAINT")
+			e.statusCode = 409;
+		else
+			e.statusCode = 500;
 		throw e;
 	}
 }
@@ -35,7 +39,7 @@ export const authRegister = async function (req, reply) {
 
 
 export const authLogin = async function (req, reply) {
-	console.log(`\n${JSON.stringify(req.body)}\n`);
+	console.log(`\nauthLogin req.body: ${JSON.stringify(req.body)}\n`);
 
 	const selectFromDB = (sql, params) => {
 		return (new Promise((resolve, reject) => {
@@ -59,17 +63,18 @@ export const authLogin = async function (req, reply) {
 
 	try {
 		const userHashedPassword = await selectFromDB('SELECT password FROM users WHERE email = ?', [req.body.email]);
-
+		if (!userHashedPassword)
+			throw httpError(401, "Invalid email or password");
 		const match = await app.bcrypt.compare(req.body.password, userHashedPassword.password);
 		if (match === true)
 		{
 			const userInfo = await selectFromDB('SELECT id, username FROM users WHERE email = ?', [req.body.email]);
-			console.log(`\n${JSON.stringify(userInfo)}\n`);
+			console.log(`\nauthLogin userInfo: ${JSON.stringify(userInfo)}\n`);
 
 			const access_tok = app.jwt.sign(userInfo , { expiresIn: '5m' });
 			const refresh_tok = app.jwt.sign(userInfo , { expiresIn: '1d' });
 
-			console.log(`\naccess_token: ${access_tok}\nrefresh_token: ${refresh_tok}\n`);
+			console.log(`\nauthLogin access_token: ${access_tok}\nauthLogin refresh_token: ${refresh_tok}\n`);
 
 			const add_to_db = await insertToken(`INSERT INTO refreshed_tokens(user_id, token)
 				VALUES (?, ?)`, [userInfo.id, refresh_tok]);
@@ -83,10 +88,9 @@ export const authLogin = async function (req, reply) {
 				.code(200)
 				.send( { access_token: access_tok }));
 		}
-		const err = new Error("Invalid email or password");
-		err.statusCode = 401;
-		throw err;
+		throw new Error("Invalid email or password").statusCode(401);
 	} catch (err) {
+		console.error(`\nERROR authLogin: ${err.message}\n`);
 		if(err.statusCode)
 			throw err;
 		err.statusCode = 500;
@@ -97,7 +101,7 @@ export const authLogin = async function (req, reply) {
 
 //le front met un header dans la req: Authorization: Bearer <token>
 export const authMe = async function (req, reply) {
-	console.log(`\n${JSON.stringify(req.user)}\n`);
+	console.log(`\nautMe req.user: ${JSON.stringify(req.user)}\n`);
 
 	const selectFromDB = (sql, params) => {
 		return (new Promise((resolve, reject) => {
@@ -111,9 +115,10 @@ export const authMe = async function (req, reply) {
 	
 	try {
 		const userInfos = await selectFromDB('SELECT id, username, email, createdAt FROM users WHERE id = ?', req.user.id);
-		console.log(`\n${JSON.stringify(userInfos)}\n`);
+		console.log(`\nauthMe userInfos: ${JSON.stringify(userInfos)}\n`);
 		return (reply.code(200).send(userInfos));
 	} catch (err) {
+		console.error(`\nERROR authMe: ${err.message}\n`);
 		const e = new Error("Error with Database");
 		e.statusCode = 500;
 		throw e;
@@ -144,14 +149,18 @@ export const authRefresh = async function (req, reply) {
 	}
 
 	try {
+		if (!req.cookies.refreshToken)
+			throw httpError(401, "Missing refresh token");
 		app.jwt.verify(req.cookies.refreshToken);
 		const old_token_in_db = await checkToken(`SELECT token FROM refreshed_tokens WHERE token = ?`, req.cookies.refreshToken);
-		console.log(`\nold token in db: ${old_token_in_db.token}\n`);
+		if (!old_token_in_db)
+			throw httpError(401, "Missing refresh token");
+		console.log(`\nauthRefresh old token in db: ${old_token_in_db.token}\n`);
 		
 		const decoded = app.jwt.decode(req.cookies.refreshToken);
 		const new_access_token = app.jwt.sign({id: decoded.id, username: decoded.username} , { expiresIn: '5m' });
 		const new_refresh_token = app.jwt.sign({id: decoded.id, username: decoded.username}, { expiresIn: '1d' });
-		console.log(`new refresh token not in db: ${new_refresh_token}\n`);
+		console.log(`authRefresh new refresh token not in db: ${new_refresh_token}\n`);
 
 		await replaceToken(`UPDATE refreshed_tokens SET token = REPLACE(token, ?, ?)`, [req.cookies.refreshToken, new_refresh_token]);
 		return (reply
@@ -165,8 +174,19 @@ export const authRefresh = async function (req, reply) {
 			.send( { access_token: new_access_token }));
 	} catch (err)
 	{
-		console.error(err.message);
-		throw err;
+		console.error(`\nERROR authRefresh: ${err.message}\n`);
+		const e = new Error();
+		if (err.name === 'TokenExpiredError' || err.name === 'JsonWebTokenError' || err.statusCode === 401)
+		{
+			e.message = "Invalid refresh token";
+			e.statusCode = 401;
+		}
+		else
+		{
+			e.message = "Internal server error"
+			e.statusCode = 500;
+		}
+		throw e;
 	}
 }
 
@@ -184,23 +204,34 @@ export const authLogout = async function (req, reply) {
 	}
 
 	try {
-		//console.log(`\n${req.cookies.refreshToken}\n`);
+		if (!req.cookies.refreshToken)
+			throw httpError(401, "Missing refresh token")
+		//console.log(`\nauthLogout req.cookies.refreshToken: ${req.cookies.refreshToken}\n`);
 		app.jwt.verify(req.cookies.refreshToken);
 		const result = await deleteFromDB(`DELETE FROM refreshed_tokens WHERE token = ?`, req.cookies.refreshToken);
 		return (reply.clearCookie('refreshToken', { path: '/' })
 		.code(204)
 		.send({ message: "User successfully logout" }));
 	} catch (err) {
-		console.error(err.message);
-		const e = new Error("Error with suppression in Database");
-		e.statusCode = 500;
+		console.error(`\nERROR authLogout: ${err.message}\n`);
+		const e = new Error();
+		if (err.name === 'TokenExpiredError' || err.name === 'JsonWebTokenError' || err.statusCode === 401)
+		{
+			e.message = "Invalid refresh token";
+			e.statusCode = 401;
+		}
+		else
+		{
+			e.message = "Internal server error"
+			e.statusCode = 500;
+		}
 		throw e;
 	}
 }
 
 
 
-
+//le front met un header dans la req: Authorization: Bearer <token>
 export const auth2faSetup = async function (req, reply) {
 	const getDbRows = (sql, params) => {
 		return (new Promise((resolve, reject) => {
@@ -224,24 +255,19 @@ export const auth2faSetup = async function (req, reply) {
 
 	try {
 		const check_in_db = await getDbRows(`SELECT twofa_enabled, twofa_secret FROM users WHERE id = ?`, [req.user.id]);
-		//console.log(`\n2faSetup check_in_db.twofa_enabled: ${check_in_db.twofa_enabled}\n2faSetup check_in_db.twofa_secret: ${check_in_db.twofa_secret}\n`)
+		//console.log(`\nauth2faSetup check_in_db.twofa_enabled: ${check_in_db.twofa_enabled}\n2faSetup check_in_db.twofa_secret: ${check_in_db.twofa_secret}\n`)
 		if (check_in_db.twofa_enabled === 1)
-		{
-			console.error("2fa already activated");
-			throw new Error();
-		}
+			throw httpError(500, "2fa already activated");
 		if (check_in_db.twofa_secret)
-		{
 			return (reply.code(201).send({ secret: check_in_db.twofa_secret }));
-		}
 
 		const secret = authenticator.generateSecret();
-		console.log(`\n2faSetup req.user: ${JSON.stringify(req.user)}\n`);
+		console.log(`\nauth2faSetup req.user: ${JSON.stringify(req.user)}\n`);
 		await updateDb(`UPDATE users SET twofa_secret = ? WHERE id = ?`, [secret, req.user.id]);
 
 		return (reply.code(201).send( { secret: secret }));
 	} catch (err) {
-		console.error(err.message);
+		console.error(`\nERROR auth2faSetup: ${err.message}\n`);
 		const e = new Error("Error with 2fa setup");
 		e.statusCode = 500;
 		throw e;
@@ -249,7 +275,7 @@ export const auth2faSetup = async function (req, reply) {
 }
 
 
-
+//le front met un header dans la req: Authorization: Bearer <token>
 export const auth2faVerify = async function (req, reply) {
 	const getSecret = (sql, params) => {
 		return (new Promise((resolve, reject) => {
@@ -273,24 +299,33 @@ export const auth2faVerify = async function (req, reply) {
 
 	try {
 		const secret = await getSecret(`SELECT twofa_enabled, twofa_secret FROM users WHERE id = ?`, [req.user.id]);
-		console.log(`\n2faVerify: secret.twofa_enabled: ${secret.twofa_enabled}\nsecret.twofa_secret: ${secret.twofa_secret}`);
+		console.log(`\nauth2faVerify: secret.twofa_enabled: ${secret.twofa_enabled}\nsecret.twofa_secret: ${secret.twofa_secret}`);
 		if (secret.twofa_enabled === 1)
-			throw new Error("2fa already enabled");
+			throw httpError(500, "2fa already enabled");
 		else if (!secret.twofa_secret)
-			throw new Error("2fa secret missing");
+			throw httpError(500, "2fa secret missing");
 		const code = req.body.code;
 
 		const isVerified = authenticator.verify({ token: code, secret: secret.twofa_secret });
-		console.log(`\n2faVerify: isverified: ${isVerified}\n`);
+		console.log(`\nauth2faVerify: isverified: ${isVerified}\n`);
 		if (isVerified === false)
-			throw new Error("token not verified");
+			throw httpError(401, "token not verified");
 		await updateDb(`UPDATE users SET twofa_enabled = ? WHERE id = ?`, [1, req.user.id]);
 
 		return (reply.code(201).send({ message: "2fa activated"}));
 	} catch (err) {
 		console.error(`\nERROR in 2faVerify: ${err.message}\n`);
 		const e = new Error("Error with 2fa verify");
-		e.statusCode = 500;
-		throw e
+		if (err.statusCode === 401)
+		{
+			e.message = "Unauthorized";
+			e.statusCode = 401;
+		}
+		else
+		{
+			e.message = "Internal server error"
+			e.statusCode = 500;
+		}
+		throw e;
 	}
 }
